Add files via upload
Detailed code for the self driving car, with required comments. Uses the servo library to control the servo motor used to turn the steering.
The two ultrasonics sensors are made to send pulses, and the time taken is returned by the output pin.
The distance from sensor to obstacle is calculated in the code using the time value output from the sensors.